requirements.txt
streamlit
groq
sentence-transformers
faiss-cpu
numpy
pypdf2
pdfplumber
python-dotenv
pytest
